Extract useUpdateSearchParam hook
